badanxia

+++ models/memo.go
@@ -1,12 +1,12 @@
 package models
 
 type MemosModel struct {
-	ID         uint   `json:"id"`
+	ID         uint64 `json:"id"`
 	Title      string `json:"title"`
 	Content    string `json:"content"`
 	CreateTime int64  `json:"create_time"`
-	UserId     uint   `json:"user_id"`
-	ClassifyId uint   `json:"classify_id"`
+	UserId     uint64 `json:"user_id"`
+	ClassifyId uint64 `json:"classify_id"`
 	UpdateTime int64  `json:"update_time"`
 	DeleteTime int64  `json:"delete_time"`
 }
