+++ models/memoClassify.go
@@ -1,11 +1,11 @@
 package models
 
 type MemoClassifyModel struct {
-	ID         uint `json:"id"`
+	ID         uint64 `json:"id"`
 	Name       string `json:"name"`
-	CreateTime int64 `json:"create_time"`
-	UserId     uint `json:"user_id"`
-	DeleteTime int64 `json:"delete_time"`
+	CreateTime int64  `json:"create_time"`
+	UserId     uint64 `json:"user_id"`
+	DeleteTime int64  `json:"delete_time"`
 }
 
 func (MemoClassifyModel) TableName() string {
